+++ CHANGELOG.md
@@ -31,6 +31,8 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 + Added `aria-label` describing mascot image (#808)
++ More reliably collapses trivial `list-of-links` widgets to `basic` widgets
+  (#827)
 
 ### Removed